Split class Image_Canvas_Tool in 2 before modyfing font handling and adapt code
We split it in a class for fonts (Image_Canvas_Font_Tools) and one for geometric
operation (Image_Canvas_Geometric_Tools).
The API of each class is not modified here so the adaptation is very simple.
Image_Canvas_Font_Tools will be modified in coming commits.

Also clean the files a bit.