Refactor font handling
The main idea is that subclasses manage the font.
This is needed because the current way doesn't work
for non-GD canvases.

To do so, Image_Canvas_Font_Tools is more flexible:
  - it can search in standard Unix and Mac folders
  - it can install fonts

Technically, Image_Canvas_Font_Tools is now a singleton
so it's more dynamic than static methods.
Also, the mapping file is not read every time.

Only GD canvas are adapted for now.
Also GitHub actions font installation mechanisms.
Last but not least, the mapping file is empty by default.